requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pytest==8.3.4
httpx==0.28.1
PyJWT==2.10.1
python-multipart==0.0.20
